prevent using CRUD routes  on user model while logged out

+++ app/controllers/users_controller.rb
@@ -2,14 +2,22 @@ class UsersController < ApplicationController
 
   # GET: /users
   get "/users" do
-    @current_user = current_user
-    @users = User.all
-    erb :"/users/index"
+    if logged_in?
+      @current_user = current_user
+      @users = User.all
+      erb :"/users/index"
+    else
+      redirect "/login"
+    end
   end
 
   # GET: /users/new. Creating a new user can be handled by logging out and signing up again.
   get "/users/new" do
-    redirect "/logout"
+    if logged_in?
+      redirect "/logout"
+    else
+      redirect "/login"
+    end
   end
 
   # POST: /users This route is not applicable to this program
@@ -19,14 +27,22 @@ class UsersController < ApplicationController
 
   # GET: /users/:id
   get "/users/:id" do
-    @user = User.find_by(id: params[:id])
-    erb :"/users/show"
+    if logged_in?
+      @user = User.find_by(id: params[:id])
+      erb :"/users/show"
+    else
+      redirect "/login"
+    end
   end
 
   # GET: /users/5/edit
   get "/users/:id/edit" do
-    @user = User.find_by(id: params[:id])
-    erb :"/users/edit"
+    if logged_in?
+      @user = User.find_by(id: params[:id])
+      erb :"/users/edit"
+    else
+      redirect "/login"
+    end
   end
 
   # PATCH: /users/5
@@ -41,9 +57,11 @@ class UsersController < ApplicationController
 
   # DELETE: /users/5/delete
   delete "/users/:id" do
-    user = User.find_by(id: params[:id])
-    user.destroy
-    session.clear
-    redirect "/"
+    if logged_in?
+      user = User.find_by(id: params[:id])
+      user.destroy
+      session.clear
+      redirect "/"
+    end
   end
 end
